# docs/50-api/openapi/sgtm-v1.yaml
...
openapi: 3.1.0
info:
  title: SGTM — Sistema de Gestion Tributaria Municipal
  version: 1.0.0-borrador
  description: |
    Contrato de la API del SGTM, derivado de las pantallas del prototipo de
    interfaz. Una operacion por opcion del menu.

    El identificador de municipalidad NO viaja en ninguna ruta, parametro ni
    cuerpo: sale del claim `municipalidad_id` del token validado (ADR-0005).
    Una peticion que lo mande por otro camino no lo consigue.

    Los importes son cadenas decimales, nunca numeros de coma flotante
    (RNF-055), y toda cifra de deuda indica a que fecha esta actualizada
    (RNF-075).
servers:
  - url: /api/v1
    description: Camino base
security:
  - tokenDeAcceso: []
...
paths:
...
  "/autorizaciones/anuncios":
    get:
      operationId: anuncios
      summary: "Anuncio y propaganda"
      description: "Autorización para instalar elementos publicitarios. La tasa resulta del área del anuncio, el número de lados y su clase."
      tags: ["Autorizaciones y licencias"]
      parameters:
        - name: nroAutorizacion
          in: query
          required: false
          description: "Filtro «Nro. Autorización» de la pantalla"
          schema: { type: string }
        - name: contribuyente
          in: query
          required: false
          description: "Filtro «Contribuyente» de la pantalla"
          schema: { type: string }
        - name: rUC
          in: query
          required: false
          description: "Filtro «R.U.C.» de la pantalla"
          schema: { type: string }
        - name: nExpediente
          in: query
          required: false
          description: "Filtro «Nº Expediente» de la pantalla"
          schema: { type: string }
        - name: direccion
          in: query
          required: false
          description: "Filtro «Dirección» de la pantalla"
          schema: { type: string }
        - name: dNI
          in: query
          required: false
          description: "Filtro «D.N.I.» de la pantalla"
          schema: { type: string }
        - name: pagina
          in: query
          required: false
          description: "Pagina que se pide, contada desde 0"
          schema: { type: string }
          example: "0"
        - name: tamano
          in: query
          required: false
          description: "Filas por pagina"
          schema: { type: string }
          example: "20"
        - name: ordenarPor
          in: query
          required: false
          description: "Campo por el que se ordena, en camelCase"
          schema: { type: string }
      responses:
        200:
          description: Operacion realizada
          content:
            application/json:
              schema: { type: object }
        "403":
          $ref: "#/components/responses/SinMunicipalidad"
        "422":
          $ref: "#/components/responses/ErrorDeValidacion"
    post:
      operationId: registrar_anuncio
      summary: "Registro de autorización de anuncio"
      description: "Registra una autorización de anuncio y GENERA SU DEUDA por la tasa (RF-114). El cuerpo lleva el titular, el establecimiento asociado —opcional—, la clase y el tipo del elemento, sus medidas, la ubicación, la vigencia y la observación del usuario, obligatoria (RNF-052). El número de la autorización lo pone el sistema desde su correlativo y la tasa sale del conjunto sellado: ninguno de los dos viene en el cuerpo. La cabecera `Idempotency-Key` se lee: reenviar el mismo registro devuelve 200 con la autorización de la primera vez y no genera un segundo cargo."
      tags: ["Autorizaciones y licencias"]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        201:
          description: Operacion realizada
          content:
            application/json:
              schema: { type: object }
        "403":
          $ref: "#/components/responses/SinMunicipalidad"
        "422":
          $ref: "#/components/responses/ErrorDeValidacion"
  "/autorizaciones/anuncios/{id}/renovacion":
    post:
      operationId: renovar_anuncio
      summary: "Renovación de autorización de anuncio"
      description: "Prorroga la autorización por otro ejercicio y devenga otra vez la tasa (RF-114). Un anuncio cesado o retirado no se renueva, y una misma autorización no devenga dos veces el mismo ejercicio. El `id` de la ruta es el número impreso de la autorización."
      tags: ["Autorizaciones y licencias"]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        201:
          description: Operacion realizada
          content:
            application/json:
              schema: { type: object }
        "403":
          $ref: "#/components/responses/SinMunicipalidad"
        "422":
          $ref: "#/components/responses/ErrorDeValidacion"
  "/autorizaciones/anuncios/{id}/cese":
    post:
      operationId: cesar_anuncio
      summary: "Cese de autorización de anuncio"
      description: "Deja sin efecto la autorización, con su motivo (RF-114). Detiene la deuda futura —un anuncio cesado no se renueva— y NO toca la ya devengada: no borra ni reversa ningún cargo (regla 4, RNF-051). El cuerpo lleva la fecha, el motivo y la observación del usuario, obligatoria (RNF-052)."
      tags: ["Autorizaciones y licencias"]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        201:
          description: Operacion realizada
          content:
            application/json:
              schema: { type: object }
        "403":
          $ref: "#/components/responses/SinMunicipalidad"
        "422":
          $ref: "#/components/responses/ErrorDeValidacion"
  "/autorizaciones/anuncios/{id}/retiro":
    post:
      operationId: retirar_anuncio
      summary: "Retiro del elemento publicitario"
      description: "Registra que el elemento se retiró de la calle, comprobado en campo (RF-114). Va después del cese: primero la autorización deja de regir y después el soporte desaparece. Al revés, el padrón diría que se desmontó un anuncio que sigue autorizado."
      tags: ["Autorizaciones y licencias"]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        201:
          description: Operacion realizada
          content:
            application/json:
              schema: { type: object }
        "403":
          $ref: "#/components/responses/SinMunicipalidad"
        "422":
          $ref: "#/components/responses/ErrorDeValidacion"
  "/autorizaciones/anuncios/reportes":
    post:
      operationId: anuncios_reportes
      summary: "Reportes de anuncio y propaganda"
      description: "Emite el padrón de autorizaciones de anuncio y propaganda por contribuyente, dirección, estado o intervalo de fechas."
      tags: ["Autorizaciones y licencias"]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        201:
          description: Operacion realizada
          content:
            application/json:
              schema: { type: object }
        "403":
          $ref: "#/components/responses/SinMunicipalidad"
        "422":
          $ref: "#/components/responses/ErrorDeValidacion"
...
  "/seguridad/accesos":
    get:
      operationId: accesos
      summary: "Accesos y políticas"
      description: "Opciones de menú y políticas del sistema controlado. La búsqueda admite filtrar por tipo y por parte del nombre del acceso."
      tags: ["Seguridad"]
      parameters:
        - name: tipo
          in: query
          required: false
          description: "Filtro «Tipo» de la pantalla"
          schema: { type: string }
        - name: nombreDelAcceso
          in: query
          required: false
          description: "Filtro «Nombre del acceso» de la pantalla"
          schema: { type: string }
        - name: pagina
          in: query
          required: false
          description: "Pagina que se pide, contada desde 0"
          schema: { type: string }
          example: "0"
        - name: tamano
          in: query
          required: false
          description: "Filas por pagina"
          schema: { type: string }
          example: "20"
        - name: ordenarPor
          in: query
          required: false
          description: "Campo por el que se ordena, en camelCase"
          schema: { type: string }
        - name: direccion
          in: query
          required: false
          description: "ASCENDENTE | DESCENDENTE"
          schema: { type: string }
          example: "ASCENDENTE"
      responses:
        200:
          description: Operacion realizada
          content:
            application/json:
              schema: { type: object }
        "403":
          $ref: "#/components/responses/SinMunicipalidad"
        "422":
          $ref: "#/components/responses/ErrorDeValidacion"
...
  "/seguridad/auditoria":
    get:
      operationId: auditoria
      summary: "Auditoría del sistema"
      description: "Bitácora de operaciones sensibles: anulaciones, extornos, bajas de deuda, cambios de valor y accesos fallidos."
      tags: ["Seguridad"]
      parameters:
        - name: ejercicio
          in: query
          required: false
          description: "Ejercicio de trabajo. Obligatorio: es la clave de particion de la bitacora"
          schema: { type: string }
          example: "2026"
        - name: usuario
          in: query
          required: false
          description: "Filtro «Usuario» de la pantalla"
          schema: { type: string }
        - name: accion
          in: query
          required: false
          description: "Filtro «Acción» de la pantalla"
          schema: { type: string }
        - name: desde
          in: query
          required: false
          description: "Filtro «Desde» de la pantalla"
          schema: { type: string }
        - name: hasta
          in: query
          required: false
          description: "Filtro «Hasta» de la pantalla"
          schema: { type: string }
        - name: pagina
          in: query
          required: false
          description: "Pagina que se pide, contada desde 0"
          schema: { type: string }
          example: "0"
        - name: tamano
          in: query
          required: false
          description: "Filas por pagina"
          schema: { type: string }
          example: "20"
        - name: ordenarPor
          in: query
          required: false
          description: "Campo por el que se ordena, en camelCase"
          schema: { type: string }
        - name: direccion
          in: query
          required: false
          description: "ASCENDENTE | DESCENDENTE"
          schema: { type: string }
          example: "ASCENDENTE"
      responses:
        200:
          description: Operacion realizada
          content:
            application/json:
              schema: { type: object }
        "403":
          $ref: "#/components/responses/SinMunicipalidad"
        "422":
          $ref: "#/components/responses/ErrorDeValidacion"
...
components:
  securitySchemes:
    tokenDeAcceso:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token OIDC validado. Debe traer el claim `municipalidad_id`; sin el, la
        peticion recibe 403 y no llega al controlador (ADR-0005, RNF-032).
  schemas:
...
    Error:
      type: object
      required: [codigo, mensaje]
      properties:
        codigo: { type: string, example: "DEUDA_YA_CANCELADA" }
        mensaje: { type: string, description: "En castellano; se muestra al usuario" }
        detalles: { type: array, items: { type: string } }
  responses:
    SinMunicipalidad:
      description: |
        El token no identifica una municipalidad. No hay valor por omision ni
        modo sin municipalidad.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ErrorDeValidacion:
      description: La peticion no cumple una regla de negocio
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }